project/wanderwise-antara-news-scraper/pyproject.toml
[tool.poetry]
name = "antara-news-scraper"
version = "0.1.0"
description = "scraper for antara news website"
authors = ["wiratana <andikawiratana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mongoengine = "^0.27.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
lxml = "^4.9.3"
deep-translator = {version = "^1.11.4", extras = ["docx", "pdf", "ai"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
